Changed SEU counter to be 16 bit

+++ main.h
@@ -12,7 +12,7 @@ volatile uint8_t CHI_UART_RX_BUFFER[CHI_UART_RX_BUFFER_SIZE];
 // CHIMERA Memory Status Structure
 struct __attribute__((packed)) CHI_Memory_Status_Str {
 	uint8_t no_SEFI_seq; // number of SEFIs in a row
-	uint8_t no_SEU;
+	uint16_t no_SEU;
 	uint8_t no_LU;
     uint8_t no_SEFI_timeout;
     uint8_t no_SEFI_wr_error;
